Move server setup functions to module
Allows to reuse the server setup functions in tests
